refactor(dot11): extract management fixed fields into fixed_fields module
